mediatek: connectivity/wlan: Fix void-pointer-to-int-cast

+++ drivers/misc/mediatek/connectivity/wlan_drv_gen4m/mgmt/scan_fsm.c
@@ -858,7 +858,7 @@ void scnEventScanDone(IN struct ADAPTER *prAdapter,
 			= prScanDone->rSparseChannel.ucChannelNum;
 		num = prScanInfo->ucSparseChannelArrayValidNum
 			= prScanDone->ucSparseChannelArrayValidNum;
-		if (prAdapter->rWifiVar.u2CountryCode != (uint16_t)NULL) {
+		if (prAdapter->rWifiVar.u2CountryCode != (long)NULL) {
 			log_dbg(SCN, INFO, "Country Code = %c%c, Detected_Channel_Num = %d\n",
 				((prAdapter->rWifiVar.u2CountryCode
 					& 0xff00) >> 8),
